new LifeCycle directory. removed redundant/commented out code.
