+++ TLM/TLM/LifeCycle/TrafficManagerMod.cs
@@ -1,4 +1,4 @@
-namespace TrafficManager {
+namespace TrafficManager.LifeCycle {
     using ColossalFramework.Globalization;
     using ColossalFramework.UI;
     using CSUtil.Commons;
