+++ TLM/TLM/State/Options.cs
@@ -8,6 +8,7 @@ namespace TrafficManager.State {
     using TrafficManager.UI.Helpers;
     using TrafficManager.UI;
     using UnityEngine;
+    using TrafficManager.LifeCycle;
 
     public class Options : MonoBehaviour {
 #if DEBUG
