+++ TLM/TLM/TLM.csproj
@@ -150,6 +150,7 @@
     <Compile Include="Custom\AI\CustomVehicleAI.cs" />
     <Compile Include="Custom\Data\CustomVehicle.cs" />
     <Compile Include="Geometry\Impl\SegmentEndId.cs" />
+    <Compile Include="LifeCycle\Patcher.cs" />
     <Compile Include="Manager\AbstractCustomManager.cs" />
     <Compile Include="Manager\AbstractFeatureManager.cs" />
     <Compile Include="Manager\AbstractGeometryObservingManager.cs" />
@@ -211,8 +212,6 @@
     <Compile Include="Patch\_NetManager\FinalizeSegmentPatch.cs" />
     <Compile Include="Patch\_VehicleManager\CreateVehiclePatch.cs" />
     <Compile Include="Patch\_VehicleManager\ReleaseVehiclePatch.cs" />
-    <Compile Include="Patch\_Vehicle\UnspawnPatch.cs" />
-    <Compile Include="Patch\_Vehicle\SpawnPatch.cs" />
     <Compile Include="State\ConfigData\Gameplay.cs" />
     <Compile Include="State\ConfigData\DebugSettings.cs" />
     <Compile Include="State\ConfigData\DynamicLaneSelection.cs" />
@@ -247,7 +246,7 @@
     <Compile Include="Custom\AI\CustomRoadBaseAI.cs" />
     <Compile Include="Custom\AI\CustomTrainAI.cs" />
     <Compile Include="Custom\PathFinding\CustomPathManager.cs" />
-    <Compile Include="LoadingExtension.cs" />
+    <Compile Include="LifeCycle\LoadingExtension.cs" />
     <Compile Include="Traffic\ExtVehicleType.cs" />
     <Compile Include="Traffic\Impl\SegmentEnd.cs" />
     <Compile Include="UI\RoadSelectionPanels.cs" />
@@ -334,8 +333,8 @@
     <Compile Include="Util\RoundaboutMassEdit.cs" />
     <Compile Include="Util\SegmentLaneTraverser.cs" />
     <Compile Include="Util\TinyDictionary.cs" />
-    <Compile Include="State\SerializableDataExtension.cs" />
-    <Compile Include="ThreadingExtension.cs" />
+    <Compile Include="LifeCycle\SerializableDataExtension.cs" />
+    <Compile Include="LifeCycle\ThreadingExtension.cs" />
     <Compile Include="TrafficLight\Impl\CustomSegment.cs" />
     <Compile Include="TrafficLight\Impl\CustomSegmentLight.cs" />
     <Compile Include="UI\ToolMode.cs" />
@@ -344,7 +343,7 @@
     <Compile Include="TrafficLight\Impl\TimedTrafficLights.cs" />
     <Compile Include="UI\TrafficManagerTool.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TrafficManagerMod.cs" />
+    <Compile Include="LifeCycle\TrafficManagerMod.cs" />
     <Compile Include="TrafficLight\Impl\TimedTrafficLightsStep.cs" />
     <Compile Include="Manager\Impl\TrafficPriorityManager.cs" />
     <Compile Include="UI\ModUI.cs" />
