+++ TLM/TLM/UI/MainMenu/MainMenuWindow.cs
@@ -16,6 +16,7 @@ namespace TrafficManager.UI.MainMenu {
     using TrafficManager.U.Panel;
     using TrafficManager.UI.MainMenu.OSD;
     using UnityEngine;
+    using TrafficManager.LifeCycle;
 
     public class MainMenuWindow
         : U.Panel.BaseUWindowPanel,
