+++ TLM/TLM/UI/ModUI.cs
@@ -5,6 +5,7 @@ namespace TrafficManager.UI {
     using TrafficManager.UI.MainMenu;
     using TrafficManager.Util;
     using UnityEngine;
+    using TrafficManager.LifeCycle;
 
     /// <summary>
     /// Globally available UI manager class which contains the main menu button and the panel.
